Added conformal remapping onto sphere in the Mathematica file
Used the stereographic projection in Mathematica to redo the Schrodinger equation on a sphere instead, whilst remapping it onto the plane. A future commit will contain the python based numerical simulation of this system with a Dirichlet boundary condition at conformal infinity (i.e. the N pole) to preserve unitarity.